refactor: painpoint Card 구현
