+++ apps/landing/src/widgets/landing/ui/hero-section/pain-point-item.tsx
@@ -1,0 +1,17 @@
+import { ReactNode } from 'react';
+
+import * as styles from './pain-point-item.css';
+
+interface PainPointItemProps {
+  icon: ReactNode;
+  label: string;
+}
+
+const PainPointItem = ({ icon, label }: PainPointItemProps) => (
+  <div className={styles.painPoint}>
+    {icon}
+    {label}
+  </div>
+);
+
+export default PainPointItem;
